Add correlation id to HTTP headers and update Pom version
The update mainly involves two significant changes. First, a correlation id is added to the HTTP headers in the NotificationSync and Core classes. This id is used for logging and tracking requests across microservices. Additionally, the version in the Pom.xml file has been updated. The latest version reflects the changes accurately.